:recycle:refactor: 데이터 패칭 함수 개선
