+++ app/(dashboard)/mydashboard/_components/invitations.tsx
@@ -1,0 +1,7 @@
+export default function Invitaions() {
+  return (
+    <section>
+      초대 받은 대시보드 리스트
+    </section>
+  );
+}
