+++ app/(dashboard)/mydashboard/page.tsx
@@ -1,3 +1,12 @@
+import DashBoards from './_components/dashboards';
+import Invitaions from './_components/invitations';
+
 export default function MyDashboard() {
-  return <>mydashboard 페이지</>;
+  return (
+    <main className="">
+      mydashboard 페이지
+      <DashBoards />
+      <Invitaions />
+    </main>
+  );
 }
